[Udpate] PlayerSetting Property

+++ MSSPlayer/Classes/PlayerSetting.swift
@@ -8,13 +8,13 @@
 import Foundation
 
 public protocol PlayerSetting {
-    var fastForwardSpeed: Double { get set }
-    var voiceAdjustSpeed: Double { get set }
-    var brightnessAdjustSpeed: Double { get set }
+    var fastForwardSpeed: Float { get set }
+    var voiceAdjustSpeed: Float { get set }
+    var brightnessAdjustSpeed: Float { get set }
 }
 
 open class MSSPlayerSetting: PlayerSetting {
-    public var fastForwardSpeed: Double = 1.0
-    public var voiceAdjustSpeed: Double = 1.0
-    public var brightnessAdjustSpeed: Double = 1.0
+    public var fastForwardSpeed: Float = 1.0
+    public var voiceAdjustSpeed: Float = 1.0
+    public var brightnessAdjustSpeed: Float = 1.0
 }
